tests: golden harness for rain_ice + shallow_convection; fix ptadjs precision
Extends the independent-oracle golden harness to the remaining two CPU
routines, so all four (ice_adjust, rain_ice, shallow_convection, turb)
are now validated against a native PHYEX call.

- rain_ice: oracle calls RAIN_ICE directly; binding is bit-identical
  (max abs diff 0.0) across all in/out + output arrays.
- shallow_convection: oracle calls SHALLOW_CONVECTION directly on a column
  tuned to actually trigger Kain-Fritsch (non-trivial golden). The cloud-
  top/base indices and the updraft mass flux PUMF are bit-identical; the
  convective tendencies are checked with a small tolerance because KF's
  iterative CAPE closure is mildly sensitive to cross-build FP
  reassociation (PUMF + indices pin the structure exactly, so coverage
  stays tight).

Bug surfaced and fixed by the harness: the Cython shallow_convection
wrapper declared `ptadjs` as a C float (single precision) while the C/F90
ABI is double. Widening it to `double` made PUMF bit-identical to the
native call. (Latent precision loss for any non-float-exact ptadjs.)

Suite: 18 passed, 1 skipped.

+++ cmake/bridge/_phyex_wrapper.pyx
@@ -527,7 +527,7 @@ def shallow_convection(
     int kbdia,
     int ktdia,
     bint osettadj,
-    float ptadjs,
+    double ptadjs,
     bint och1conv,
     int kch1,
     # 1D input arrays (Fortran-contiguous)
